requirements.txt
alembic==1.11.1
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.101.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
joblib==1.3.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==1.26.4
pandas==2.2.2
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.3.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.14
starlette==0.27.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.23.2
watchfiles==1.1.1
websockets==15.0.1
